# openapi.yaml
openapi: 3.0.0
info:
  title: NBGN Voucher Backend API
  description: Backend API for NBGN voucher system on Arbitrum One
  version: 1.0.0
  contact:
    name: Molly (Backend Engineer)
    email: molly@nbgn.io
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.nbgn.io
    description: Production server
...
paths:
...
  /api/users/{address}:
    get:
      tags: [Users]
      summary: Get user profile
      description: Retrieve user profile with transaction statistics
      operationId: getUserProfile
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
            pattern: '^0x[a-fA-F0-9]{40}$'
          description: Ethereum address
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User not found
  /api/users/username:
    post:
      tags: [Users]
      summary: Set username
      description: Set or update username with signature verification. Rate limited to 5 per hour.
      operationId: setUsername
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUsernameRequest'
      responses:
        '200':
          description: Username set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Username already taken
        '401':
          description: Invalid signature
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /api/transactions/{address}:
    get:
      tags: [Users]
      summary: Get user transactions
      description: Get paginated list of user's NBGN transactions
      operationId: getUserTransactions
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
            pattern: '^0x[a-fA-F0-9]{40}$'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  total:
                    type: integer
                    example: 42
                  limit:
                    type: integer
                    example: 50
                  offset:
                    type: integer
                    example: 0
  /api/transactions/recent:
    get:
      tags: [Users]
      summary: Get recent transactions
      description: Get recent NBGN transactions across all users
      operationId: getRecentTransactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Recent transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        address:
          type: string
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f2bD7E"
        username:
          type: string
          nullable: true
          example: "alice.eth"
        created_at:
          type: string
          format: date-time
        total_minted:
          type: string
          example: "1000000000000000000000"
        total_redeemed:
          type: string
          example: "500000000000000000000"
        total_burned:
          type: string
          example: "100000000000000000000"
        transaction_count:
          type: integer
          example: 42
    SetUsernameRequest:
      type: object
      required:
        - address
        - username
        - message
        - signature
      properties:
        address:
          type: string
          pattern: '^0x[a-fA-F0-9]{40}$'
        username:
          type: string
          minLength: 3
          maxLength: 20
        message:
          type: string
          description: Message that was signed
        signature:
          type: string
          description: Ethereum signature
    Transaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tx_hash:
          type: string
        block_number:
          type: integer
          format: int64
        timestamp:
          type: string
          format: date-time
        user_address:
          type: string
        transaction_type:
          type: string
          enum: [mint, redeem, burn]
        eure_amount:
          type: string
          nullable: true
        nbgn_amount:
          type: string
        gas_used:
          type: string
          nullable: true
        gas_price:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
  responses:
...
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per window
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when window resets
        Retry-After:
          schema:
            type: integer
          description: Seconds until next request allowed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Too Many Requests"
              message:
                type: string
                example: "Rate limit exceeded. Maximum 10 requests per 3600 seconds allowed."
              retry_after:
                type: integer
                example: 2547
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits
security:
  - {}  # Most endpoints don't require auth
  - ApiKeyAuth: []  # Optional for higher rate limits